CMake: Enable Introspection on Visual Studio builds
The standard windows cmd.exe does not support shebang lines, so we must
call g-ir-scanner with python.  This assumes that the correct Python is in
the path as this requires the same Python series that was used to build
_giscanner.pyd to be used.  Possibly improve by using a PYTHON envvar or
detect a Python installation and make it adjustable in case multiple
Python installations exist on the system.