CMake: Use perl to call glib-mkenums for Visual Studio builds
The standard cmd.exe that is used by the NMake Makefiles and Visual Studio
projects do not support shebang lines, so we need to call glib-mkenums
using PERL explicitly.  This assumes a PERL interpretor is in the PATH--we
can improve on this by detecting a PERL installation and use that instead.