glib: Annotate public symbols with export macros
For Windows/Visual Studio builds, mark the public symbols in the headers
with export macros so that we may export them during the build and import
them when using these symbols, like what we do for the CPP, QT4 and QT5
interfaces of poppler.  For other builds, these macros expand to nothing,
so things remain as they were.